Add RGL's NumeralDut.gf with a tiny modification
This was needed to get generalized quantifiers (e.g. `at least`, `exactly`)
to work in Dutch.

The modification was to restrict the inheritance of CatDut to [Numeral,Digits].
A similar restriction is done in RGL's NumeralGer. Before, this restriction was
applied in AttemptoDut to NumeralDut but in this case the Card-category will
not be present which results in empty linearizations.

Hopefully the modified NumeralDut.gf can be pushed into the RGL because it
really shouldn't be part of ACE-in-GF.