Achieved basic wh in object position Parsed: 17523 (90.22%) - (94.56% unambiguous) NOT parsed: 1899 (9.77%)
I still needed to add two abstract rules whobject_QS and neg_whobject_QS,
but not sure what linearisation these should have in other languages.
In addition their implementation is probably flawed, because it
fails if the parameter is not OQuest.
First alternative is to consider moving these to a non-common abstract module,
e.g. AttemptoQuestionsAceAbs or something.
Another is to move the logic to the highest level (dealing with S/SimpleS).
This would probably give coverage of a bunch of other such as question words
as teh objects of relative clauses.