Add tools to linearize the Codeco subset
into all the languages. This demonstrates the performance
of the system in linearizing (9 minutes on my laptop). Also
it revealed that VPCoord is not the only construct that some
languages cannot deal with. E.g. Finnish often (partially)
fails and Fre and Ita sometimes produce empty output.
I didn't study the missing linearizations closely though.