Add isNeg=False to lin ipNPQ ip = lin NP {...}
This is required by the latest RGL which adds the isNeg feature
to the NP. heavyNP has this feature set to False. So, now that
I set it always to False in questions (ipNPQ) makes some of the
questions wrong (but I guess they were already wrong before,
hard to know because the questions are not under regression test).

Some refactoring is needed to combine the ipNPQ of the Romance
languages.

Ontograph lins are also updated, e.g. German is better now.