Geography: add languages: Dut, Fin, Ita, Spa
The lexicon files are automatically generated from the
Google Drive spreadsheet

https://docs.google.com/spreadsheet/ccc?key=0Aprlzd4hAZrEdGVmYnBwZWdyVkUwQ0RBZE5BSXB1OEE

The words in the new languages have been Google-translated from the English
words in the first column and used in 1-arg smart paradigms.
So many words are wrong or have wrong wordforms.
E.g. for Finnish Google Translate often fails to give the nominative case.