ACE: remove comma from numbers
Previously the grammar forced numbers to be written as

    1 , 2 3 4 , 5 6 7 , 9 0 1

which is not ACE + the comma is confusing especially in the context
of look-ahead editing.

The fix made a copy of NumeralEng and commented out commaIf.
This should be cleaned up in the future, currently
didn't get the simple extension of NumeralEng to compile.